move timing wrapper to a utility folder
