+++ ssacc/adapters/shell_controller.py
@@ -24,8 +24,8 @@ import sys
 from ssacc.clean_df import CleanDF
 from ssacc.map_ssa_zip_fips import MapSsaZipFips
 from ssacc.ssa_fips import SsaFips
-from ssacc.timing_wrapper import timing
 from ssacc.validate_map import ValidateMap
+from ssacc.wrappers.timing_wrapper import timing
 from ssacc.zip_fips import ZipFips
 
 # Todo: can we eliminate this abspath cruft?
@@ -68,9 +68,10 @@ def shell(args):
     Build verification SSA CC to ZIP CSV.
     Build refined SSA CC (3 and 5 digit) to ZIP CSV.
     """
+    args = parse_args(args)
 
     project_root = Path(__file__).resolve().parents[2]  # was 1
-    args = parse_args(args)
+
     """
     Build CSV of SSA and FIPS codes
     """
@@ -79,12 +80,13 @@ def shell(args):
     print(df_ssa_fips.columns.values)
     print(file_path)
     print(df_ssa_fips.head())
-    # ZIPs and FIPS.
 
+    # ZIPs and FIPS.
     # Build a new ZIP and FIPS CSV if asked.
+    # Takes about 5 minutes locally
     zip_fips = ZipFips()
     if args.r:
-        project_root = Path(__file__).parents[1]  # was 2
+        project_root = Path(__file__).parents[2]  # was 2
         file_path = project_root.joinpath("data", "source")
         print(f"root file path {file_path}")
         zip_fips.files_to_csv(file_path)
@@ -111,6 +113,7 @@ def shell(args):
     df_map_2 = df_map_2.sort_values(by=["zip", "ssacnty"])
     file_path = project_root.joinpath("data", "temp", "ssa_zip_fips.csv")
     df_map_result = MapSsaZipFips.write_csv(df_map_2, file_path)
+
     # Validate the ZIP SSA table.
     result = ValidateMap.validate(file_path)
     if not result:
