+++ ssacc/map_ssa_zip_fips.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import pandas as pd
 
 from ssacc.clean_df import CleanDF
-from ssacc.timing_wrapper import timing
+from ssacc.wrappers.timing_wrapper import timing
 
 print("Running" if __name__ == "__main__" else "Importing", Path(__file__).resolve())
