+++ ssacc/use_cases/README.txt
@@ -1,0 +1,10 @@
+These seem like use cases
+
+    """
+    Build SSA CC to FIPS CC mapping.
+    Build ZIP to FIPS CC mapping.
+    Build joined ZIP, FIPS CC, SSA CC.
+    Add city, state.
+    Build verification SSA CC to ZIP CSV.
+    Build refined SSA CC (3 and 5 digit) to ZIP CSV.
+    """
