+++ ssacc/validate_map.py
@@ -7,7 +7,7 @@ import numpy as np
 import pandas as pd
 from pandas.io.parsers import ParserError
 
-from ssacc.timing_wrapper import timing
+from ssacc.wrappers.timing_wrapper import timing
 
 print("Running" if __name__ == "__main__" else "Importing", Path(__file__).resolve())
